Added Proper GridLevel controller
The GridLevel game object is responsible for controlling the level grid. It can update and keep track of how many Units are on a certain GridObject or at its position. Currently will break when a character runs outside of the grid, but that will be fixed naturally via validating player inputs when choosing where the character should move to.